// src/types.ts
export interface Server {
  id: number;
  name: string;
  url: string;
  downloadUrl: string;
  isActive: boolean;
}

export interface Episode {
  id: number;
  title?: string;
  thumbnail?: string; 
  description?: string; 
  duration?: string; 
  progress: number; 
  servers: Server[];
  isLastEpisode?: boolean;
  badgeText?: string; 
  // حقول الجدولة الجديدة
  isScheduled?: boolean;
  scheduledAt?: string;
  notifyOnPublish?: boolean;
  virtual_path?: string;
  telegramOriginalUrl?: string;
}